# swagger.json
{"openapi":"3.0.0","info":{"title":"Todo example","description":"The Todo API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateCatDto":{"type":"object","properties":{}},"UpdateCatDto":{"type":"object","properties":{}},"CreateControllerForGuardDto":{"type":"object","properties":{}},"UpdateControllerForGuardDto":{"type":"object","properties":{}},"TodoEntity":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","description","createdAt","updatedAt"]},"GetTodoListResponse":{"type":"object","properties":{"todoes":{"type":"array","items":{"$ref":"#/components/schemas/TodoEntity"}}},"required":["todoes"]},"CreateTodoDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title","description"]},"UpdateTodoDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title","description"]}}},"paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/cats":{"post":{"operationId":"CatsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCatDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"CatsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/cats/{id}":{"get":{"operationId":"CatsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"CatsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCatDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"CatsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/controller-for-guard":{"post":{"operationId":"ControllerForGuardController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateControllerForGuardDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"ControllerForGuardController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/controller-for-guard/{id}":{"get":{"operationId":"ControllerForGuardController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"ControllerForGuardController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateControllerForGuardDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ControllerForGuardController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/todos":{"get":{"operationId":"TodoController_index","parameters":[],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetTodoListResponse"}}}}},"tags":["todo"]},"post":{"operationId":"TodoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoDto"}}}},"responses":{"201":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoEntity"}}}},"400":{"description":"Request Body is invalid"}},"tags":["todo"]}},"/todos/{todoId}":{"get":{"operationId":"TodoController_show","parameters":[{"name":"todoId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoEntity"}}}},"404":{"description":"Todo has is not found."}},"tags":["todo"]},"put":{"operationId":"TodoController_update","parameters":[{"name":"todoId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoDto"}}}},"responses":{"201":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoEntity"}}}},"400":{"description":"Request Body is invalid"},"404":{"description":"Todo has is not found."}},"tags":["todo"]},"delete":{"operationId":"TodoController_delete","parameters":[{"name":"todoId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoEntity"}}}},"404":{"description":"Todo has is not found."}},"tags":["todo"]}}}}